rusnas-license-server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
aiosqlite>=0.20
cryptography>=42.0
pydantic>=2.0
python-dotenv>=1.0
slowapi>=0.1.9
jinja2>=3.1
python-multipart>=0.0.9
